requirements.txt
pulp
numpy as np
matplotlib.pyplot as plt